Improvements for counters.

+++ Vostok.Metrics/Primitives/Counter/ICounter.cs
@@ -1,22 +1,26 @@
+using System;
 using JetBrains.Annotations;
+using Vostok.Metrics.Model;
 
 namespace Vostok.Metrics.Primitives.Counter
 {
     /// <summary>
     /// <para>
-    /// Counter represents a value that can only increase.
-    /// Counters with the same tags are summed up server side.
+    /// Counter represents a value that can only increase. Counters with the same tags are summed up server side.
+    /// </para>
+    /// <para>
+    /// Counter immediately produces a <see cref="MetricEvent"/> on each <see cref="Add"/> call: there's no client side aggregation.
     /// </para>
     /// <remarks>
     /// <para>
-    /// To create a Counter use <see cref="MetricContextExtensionsCounter">extensions</see> for <see cref="IMetricContext"/>.
+    /// To create a Counter, use <see cref="CounterFactoryExtensions">extensions</see> for <see cref="IMetricContext"/>.
     /// </para>
     /// </remarks>
     /// <example>
     /// <para>
-    /// You can use a Counter to represent the number of newly-registered users to your app.
-    /// Every replica of your API creates a Counter with name "registered-users".
-    /// When API replica registers new user it calls <see cref="ICounterExtensions.Inc"/>.
+    /// You can use a Counter to represent the number of newly-registered users in your app.
+    /// Every replica of your API creates a Counter named "registered-users".
+    /// When API replica registers a new user, it calls <see cref="ICounterExtensions.Increment"/>.
     /// Values from all replicas are summed up and the single value is saved to a permanent storage.
     /// </para>
     /// </example>
@@ -24,6 +28,10 @@ namespace Vostok.Metrics.Primitives.Counter
     [PublicAPI]
     public interface ICounter
     {
-        void Add(double value);
+        /// <summary>
+        /// Adds given non-negative <paramref name="value"/> to the counter.
+        /// </summary>
+        /// <exception cref="ArgumentOutOfRangeException">Passed <paramref name="value"/> was negative.</exception>
+        void Add(long value);
     }
 }
